server: add tcp ping responder, update option parser
server.c:
- Remove protocol enum, use sock_domain instead.
- Update Unix socket argument parser. Initialize unix_sock_path only
there.
- Add --listen|-L options, and update usage message correspondingly.
- Add Inet socket argument parser.
- Add check for conflicting protocols in parse_options.
- Redirect error output prints to stderr.
- Add socket domain switch to main. Since socket creation and binding
process is different for inet and unix sockets, the switch is used.
- Move unix socket creation, binding, and address structure
initialization to AF_UNIX case of domain switch.
- Add inet socket creation, binding, and address selection to AF_INET
case of domain switch. Address selection is performed with the help of
getaddrinfo, and the process is as described in getaddrinfo manpage.
- Replace read with recv, and write with send for operations on sockets.
This is useful in case we want to set flags. However, no flags are used.
- Unlink unix_sock_path on exit only if it was set beforehand.

pingpong.h.in:
- Add defines for default inet socket address and port.

+++ src/server.c
@@ -15,6 +15,10 @@
 #include <fcntl.h>
 #include <getopt.h>
 
+#include <netdb.h>
+#include <netinet/in.h>
+#include <netinet/tcp.h>
+
 #include "pingpong.h"
 
 
@@ -22,30 +26,29 @@
 
 #define LISTEN_BACKLOG 5
 
-static char *short_options = "hvl:u::";
+static char *short_options = "l:u::L::hv";
 static struct option long_options[] = {
     {"log",     required_argument, 0, 'l'},
     {"unix",    optional_argument, 0, 'u'},
+    {"listen",  optional_argument, 0, 'L'},
     {"help",    no_argument,       0, 'h'},
     {"version", no_argument,       0, 'v'},
     {0,         0,                 0, 0}
 };
 
-enum protocol
-{
-    NONE,
-    UNIX,
-};
+static int sock_domain, sock_type;
+static char unix_sock_path[108];
+static char hostname[254];
+static char service[16];
 
-static enum protocol proto = NONE;
-
-static char unix_sock_path[108] = UNIX_SOCKET_PATH;
 static char *log_path = "log.txt";
 
 static void print_version (void);
 static void print_usage (void);
 static void parse_options (int argc, char **argv);
-static void parse_unix_socket_param (const char *optarg);
+static void parse_unix_sock_addr (const char *optarg);
+static void parse_inet_sock_addr (const char *optarg);
+
 int main (int argc, char **argv);
 
 
@@ -62,29 +65,53 @@ print_usage (void)
     printf("%s: usage \n./%s [-hvu] [-l log_path] [-u sock_path]\n",
            PROGRAM_NAME, PROGRAM_NAME);
     puts(("\
-\t-h, --help              show this message\n\
-\t-u, --unix [pathname]   use Unix socket, bind it to pathname\n\
-\t-l, --log [pathname]    write log to pathname\n\
-\t-v, --version           show version"));
+\t-h, --help               show this message\n\
+\t-u, --unix [pathname]    use Unix socket, bind it to pathname\n\
+\t-L, --listen [addr:port] use TCP socket, bind it to addr:port\n\
+\t-l, --log [pathname]     write log to pathname\n\
+\t-v, --version            show version"));
     return;
 }
 
-
 static void
-parse_unix_socket_param (const char *optarg)
+parse_unix_sock_addr (const char *optarg)
 {
     char *env_var;
 
-    if (! optarg) {
-        if ((env_var = getenv("UNIX_SOCKET")) != NULL) {
-            /* Use env var if set */
-            strcpy(unix_sock_path, env_var);
-        }
-        return;
+    /* Choose a path to bind to */
+    if (optarg) {
+        strcpy(unix_sock_path, optarg);
+    //} else if ((char *env_var = getenv("UNIX_SOCKET")) != NULL) {
+    } else if ((env_var = getenv("UNIX_SOCKET")) != NULL) {
+        /* Use env var if set */
+        strcpy(unix_sock_path, env_var);
+    } else {
+        strcpy(unix_sock_path, UNIX_SOCKET_PATH);
     }
 
-    strcpy(unix_sock_path, optarg);
     return;
+}
+
+static void
+parse_inet_sock_addr (const char *optarg)
+{
+    if (optarg) {
+        /* Locate last colon ":" in the optarg */
+        char *colon_loc = strrchr(optarg, ':');
+        
+        if (colon_loc != NULL) {
+            /* colon is present, assuming arg is in host:port format */
+            strncpy(hostname, optarg, colon_loc-optarg);
+            strcpy(service, colon_loc+1);
+        } else {
+            /* colon is not present */
+            fputs("parse_inet_sockaddr: use host:port format", stderr);
+            exit(EXIT_FAILURE);
+        }
+    } else {
+        strcpy(hostname, INET_SOCKET_HOST);
+        strcpy(service, INET_SOCKET_PORT);
+    }
 
 }
 
@@ -107,6 +134,11 @@ parse_options (int argc, char **argv)
                 log_path = optarg;
                 break;
             case 'u':
+                if (sock_domain || sock_type) {
+                    fprintf(stderr, "%s: conficting options specified\n", argv[0]);
+                    goto exit_failure;
+                }
+
                 if (!optarg && NULL != argv[optind] && '-' != argv[optind][0]) {
                     /* This construct is required to pass optarg
                      * as -u path or --unix path, instead of
@@ -120,8 +152,29 @@ parse_options (int argc, char **argv)
                      */
                     temp_arg = argv[optind++];
                 }
-                parse_unix_socket_param(temp_arg);
-                proto = UNIX;
+
+                sock_domain = AF_UNIX;
+                sock_type = SOCK_STREAM;
+
+                parse_unix_sock_addr(temp_arg);
+
+                break;
+            case 'L':
+                if (sock_domain || sock_type) {
+                    fprintf(stderr, "%s: conficting options specified\n", argv[0]);
+                    goto exit_failure;
+                }
+
+                if (!optarg && NULL != argv[optind] && '-' != argv[optind][0]) {
+                    /* For the explanation, see the comment in the 'u' case */
+                    temp_arg = argv[optind++];
+                }
+
+                sock_domain = AF_INET;
+                sock_type = SOCK_STREAM;
+
+                parse_inet_sock_addr(temp_arg);
+
                 break;
             default:
                 goto exit_failure;
@@ -131,7 +184,7 @@ parse_options (int argc, char **argv)
     return;
 
 exit_failure:
-    printf("Try './%s -h' for more information.\n", PROGRAM_NAME);
+    fprintf(stderr, "Try './%s -h' for more information.\n", PROGRAM_NAME);
     exit(EXIT_FAILURE);
 }
 
@@ -139,7 +192,7 @@ int
 main (int argc, char **argv)
 {
     int sfd, cfd, log_fd;
-    struct sockaddr_un my_addr;
+    struct addrinfo *rp;
 
     ssize_t read_cnt, write_cnt, write_cum;
     char recv_buf[BUF_SIZE];
@@ -147,30 +200,70 @@ main (int argc, char **argv)
 
     parse_options(argc, argv);
 
-    if (proto == NONE) {
-        printf("No protocol specified!\n");
-        exit(EXIT_FAILURE);
-    }
-
     /* Open log file */
     log_fd = open(log_path, O_RDWR | O_CREAT | O_APPEND, S_IRWXU);
     if (log_fd == -1)
         handle_error("open log");
 
-    /* Create Unix socket */
-    sfd = socket(AF_UNIX, SOCK_STREAM, 0);
-    if (sfd == -1)
-        handle_error("socket");
+    struct sockaddr_un unix_addr;
 
-    /* Initialize my_addr structure with pathname set to unix_sock_path */
-    memset(&my_addr, 0, sizeof(struct sockaddr_un)); /* clear structure */
-    my_addr.sun_family = AF_UNIX;
-    strncpy(my_addr.sun_path, unix_sock_path, sizeof(my_addr.sun_path)-1);
+    int err;
+    struct addrinfo *inet_result;
+    struct addrinfo hints;
 
-    /* Bind socket to the address */
-    unlink(my_addr.sun_path);
-    if (bind(sfd, (struct sockaddr *) &my_addr, sizeof(struct sockaddr_un)) == -1)
-        handle_error("bind");
+    /* Create and bind socket to the address */
+    switch (sock_domain) {
+        case AF_UNIX:
+            /* Initialize unix_addr structure */
+            memset(&unix_addr, 0, sizeof(struct sockaddr_un));
+            unix_addr.sun_family = AF_UNIX;
+            strncpy(unix_addr.sun_path, unix_sock_path, sizeof(unix_addr.sun_path)-1);
+
+            unlink(unix_addr.sun_path);
+
+            sfd = socket(sock_domain, sock_type, 0);
+            if (sfd == -1)
+                handle_error("socket");
+
+            if (bind(sfd, (struct sockaddr *) &unix_addr, sizeof(struct sockaddr_un)) == -1)
+                handle_error("bind");
+
+            break;
+        case AF_INET:
+
+            memset(&hints, 0, sizeof(struct addrinfo));
+            hints.ai_family = AF_UNSPEC;
+            hints.ai_socktype = sock_type;
+            hints.ai_protocol = 0;
+            hints.ai_canonname = NULL;
+            hints.ai_addr = NULL;
+            hints.ai_next = NULL;
+
+            if ((err=getaddrinfo(hostname, service, &hints, &inet_result)) != 0) {
+                fprintf(stderr, "getaddrinfo: %s\n", gai_strerror(err));
+                exit(EXIT_FAILURE);
+            }
+
+            for (rp = inet_result; rp != NULL; rp = rp->ai_next) {
+                sfd = socket(rp->ai_family, rp->ai_socktype, rp->ai_protocol);
+                if (sfd == -1)
+                    continue;
+
+                if (bind(sfd, rp->ai_addr, rp->ai_addrlen) == 0)
+                    break;
+
+                close(sfd);
+            }
+
+            if (rp == NULL)
+                handle_error("bind");
+            break;
+        default:
+            fputs(("No protocol specified!\n"), stderr);
+            exit(EXIT_FAILURE);
+    }
+
+    freeaddrinfo(inet_result);
 
     /* Mark socket as passive socket */
     if (listen(sfd, LISTEN_BACKLOG) == -1)
@@ -183,9 +276,9 @@ main (int argc, char **argv)
 
     do {
         /* Read the message to recv_buf */
-        read_cnt=read(cfd, recv_buf, BUF_SIZE);
+        read_cnt=recv(cfd, recv_buf, BUF_SIZE, 0);
         if (read_cnt == -1)
-            handle_error("read");
+            handle_error("recv");
 
         /* Check if there is unread data in the buffer */
         if (ioctl(cfd, FIONREAD, &read_cnt) == -1)
@@ -196,19 +289,18 @@ main (int argc, char **argv)
     printf("Received %s\n", recv_buf);
 
     write_cnt = write_cum = 0;
-
     while ((write_cnt=write(log_fd, recv_buf+write_cum, strlen(recv_buf)-write_cum)) < strlen(recv_buf)-write_cum) {
         if (write_cnt == -1)
             handle_error("log write");
         write_cum += write_cnt;
     }
 
-    write_cnt = write_cum = 0;
 
     /* Check if message is PING, and reply with PONG */
+    write_cnt = write_cum = 0;
     if (strncmp(PING, recv_buf, strlen(PING)) == 0) {
         printf("Sending %s\n", send_buf);
-        while ((write_cnt=write(cfd, send_buf+write_cum, strlen(send_buf)-write_cum)) < strlen(send_buf)-write_cum) {
+        while ((write_cnt=send(cfd, send_buf+write_cum, strlen(send_buf)-write_cum, 0)) < strlen(send_buf)-write_cum) {
             if (write_cnt == -1)
                 handle_error("write");
             write_cum += write_cnt;
@@ -221,8 +313,10 @@ main (int argc, char **argv)
     if (close(cfd) == -1)
         handle_error("close cfd");
 
-    if (unlink(unix_sock_path) == -1)
-        handle_error("unlink");
+    if (strlen(unix_sock_path) > 0) {
+        if (unlink(unix_sock_path) == -1)
+            handle_error("unlink");
+    }
 
     if (close(sfd) == -1)
         handle_error("close sfd");
